Update decal menu
- Move CreateDecal() to Editor folder (was breaking the standalone
build) in a DecalMmenuItems
- New menu to create decal is: GameObject/Render Pipeline/High
Definition/DecalProjector
- Add a null check test iin AddDecal on material
