+++ ScriptableRenderPipeline/HDRenderPipeline/Decal/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d48c34658396f5e478714620e1290cfc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
